Updated dlib

+++ Android_app/app/dlib/dlib/dnn/trainer.h
@@ -14,7 +14,7 @@
 
 #include "../pipe.h"
 #include "../threads.h"
-#include "cuda_dlib.h"
+#include "../cuda/cuda_dlib.h"
 #include "../statistics/running_gradient.h"
 #include <atomic>
 #include <cstdio>
@@ -593,11 +593,8 @@ namespace dlib
 
         void record_loss(double loss)
         {
-            // Say that we will check if the gradient is bad 200 times during each
-            // iter_without_progress_thresh interval of network updates.   This kind of
-            // budgeting causes our gradient checking to use a fixed amount of
-            // computational resources, regardless of the size of
-            // iter_without_progress_thresh.
+            // This kind of budgeting causes our gradient checking to use a fixed amount of
+            // computational resources, regardless of the size of iter_without_progress_thresh.
             gradient_check_budget += 200;
 
             rs.add(loss);
@@ -705,7 +702,7 @@ namespace dlib
                                 test_steps_without_progress = 0;
                                 // Empty out some of the previous loss values so that test_steps_without_progress 
                                 // will decrease below test_iter_without_progress_thresh.  
-                                for (int cnt = 0; cnt < test_previous_loss_values_dump_amount+test_iter_without_progress_thresh/10 && test_previous_loss_values.size() > 0; ++cnt)
+                                for (unsigned long cnt = 0; cnt < test_previous_loss_values_dump_amount+test_iter_without_progress_thresh/10 && test_previous_loss_values.size() > 0; ++cnt)
                                     test_previous_loss_values.pop_front();
                             }
                         }
@@ -825,7 +822,7 @@ namespace dlib
                             steps_without_progress = 0;
                             // Empty out some of the previous loss values so that steps_without_progress 
                             // will decrease below iter_without_progress_thresh.  
-                            for (int cnt = 0; cnt < previous_loss_values_dump_amount+iter_without_progress_thresh/10 && previous_loss_values.size() > 0; ++cnt)
+                            for (unsigned long cnt = 0; cnt < previous_loss_values_dump_amount+iter_without_progress_thresh/10 && previous_loss_values.size() > 0; ++cnt)
                                 previous_loss_values.pop_front();
                         }
                     }
@@ -1281,8 +1278,8 @@ namespace dlib
         std::atomic<bool> updated_net_since_last_sync;
 
         bool sync_file_reloaded;
-        int previous_loss_values_dump_amount;
-        int test_previous_loss_values_dump_amount;
+        unsigned long previous_loss_values_dump_amount;
+        unsigned long test_previous_loss_values_dump_amount;
     };
 
 // ----------------------------------------------------------------------------------------
